requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.14.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
kafka-python==2.0.2
    # via -r requirements.in
numpy==1.24.3
    # via
    #   -r requirements.in
    #   pandas
pandas==2.0.1
    # via -r requirements.in
py4j==0.10.9.5
    # via pyspark
pyspark==3.3.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via
    #   django
    #   djangorestframework
    #   pandas
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via pandas
